+++ processor/src/test/fixtures/expected/com/example/nested/NestedRouter_InnerRouterService.java
@@ -1,7 +1,7 @@
 package com.example.nested;
 
-import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.processing.Generated;
 import router.fu.Location;
 
 @Generated("router.fu.processor.RouterProcessor")